pedometer: improve migration of active time
Default to 2 second per step but still inaccurate but better than treating it as the full bucket time. This fixes a bug where if there are new steps logged to the latest bucket after migration it would count all steps as being in a short active time period.

Issue: #3890

+++ app/src/main/java/com/kylecorry/trail_sense/tools/pedometer/domain/ActiveTimeCalculator.kt
@@ -4,7 +4,7 @@ import java.time.Duration
 
 class ActiveTimeCalculator {
 
-    fun calculate(steps: Int, elapsedTime: Duration): Duration {
+    fun calculate(steps: Long, elapsedTime: Duration): Duration {
         if (steps <= 0 || elapsedTime.isZero || elapsedTime.isNegative) {
             return Duration.ZERO
         }
